requirements.txt
scipy>=0.18.1
numpy>=1.11.2
pytest>=3.0
pandas>=0.25.3
matplotlib>=3.1.2
umap_learn>=0.3.10
seaborn>=0.10.0